ETL_job/requirements.txt
pandas
numpy
requests
fastapi
boto3
python-dotenv
openai
uvicorn
supabase
pydantic
light_embed
pydantic-settings